requirements.txt
annotated-types==0.7.0
blinker==1.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
Flask==3.1.1
flask-cors==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
language_tool_python==2.9.4
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
numpy==2.3.1
opencv-python==4.7.0.68
packaging==25.0
Pillow==9.4.0
preshed==3.0.10
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyenchant==3.2.2
Pygments==2.19.2
pyspellchecker==0.8.3
PyYAML==6.0
requests==2.32.4
rich==14.0.0
scipy==1.10.0
shellingham==1.5.4
sketchpy==0.1.0
smart_open==7.3.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
svg.path==6.2
svgpathtools==1.5.1
svgwrite==1.4.3
thinc==8.3.6
toml==0.10.2
tqdm==4.64.1
turtle==0.0.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
